# novify-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Novify API",
        "description": "Digital Marketplace Platform API Documentation",
        "contact": {
            "name": "Novify Support",
            "email": "support@novify.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://novify.solvertech.co",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/v1/bills/billers": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get billers list",
                "description": "Get list of billers with optional category filtering",
                "operationId": "e2b65144c87de6e9ba6e0a4ec8f5f410",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by bill category ID",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Billers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Electric Company"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "ELEC001"
                                                    },
                                                    "logo": {
                                                        "type": "string",
                                                        "example": "logo_url"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "biller_items": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bills/billers/{biller}": {
            "get": {
                "tags": [
                    "Bills"
                ],
                "summary": "Get specific biller",
                "description": "Get details of a specific biller with its items",
                "operationId": "0d014a83ddfab86c149dd44e6c94d6ce",
                "parameters": [
                    {
                        "name": "biller",
                        "in": "path",
                        "description": "Biller ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Biller retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Electric Company"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "ELEC001"
                                                },
                                                "logo": {
                                                    "type": "string",
                                                    "example": "logo_url"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "biller_items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/lookup/currencies": {
            "get": {
                "tags": [
                    "Lookup"
                ],
                "summary": "Get currencies list",
                "description": "Get list of active currencies",
                "operationId": "7bc440d2fa19db91d8a2a8b945b1699a",
                "responses": {
                    "200": {
                        "description": "Currencies retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "US Dollar"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "USD"
                                                    },
                                                    "symbol": {
                                                        "type": "string",
                                                        "example": "$"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "JWT Bearer Token Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "jwt": []
        }
    ]
}